requirements.txt
streamlit
pydantic
pandas
openpyxl
SQLAlchemy
langchain
langgraph
langchain-openai
python-dotenv